# x86-64 48-bit pointer optimization for GenericValue