Fix #330 

In x86-64 platform, `RAPIDJSON_48BITPOINTER_OPTIMIZATION` is turned on by default.
This reduces the `sizeof(GenericValue)` from 24 bytes to 16 bytes.
It is interesting to show that, although there is additional calculation for those pointers, the performance does improved due to reduce memory size.

~~~
With RAPIDJSON_48BITPOINTER_OPTIMIZATION=0

[ RUN      ] RapidJson.DocumentParseInsitu_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParseInsitu_MemoryPoolAllocator_SSE42 (403 ms)
[ RUN      ] RapidJson.DocumentParseIterativeInsitu_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParseIterativeInsitu_MemoryPoolAllocator_SSE42 (433 ms)
[ RUN      ] RapidJson.DocumentParse_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParse_MemoryPoolAllocator_SSE42 (457 ms)
[ RUN      ] RapidJson.DocumentParseIterative_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParseIterative_MemoryPoolAllocator_SSE42 (492 ms)
[ RUN      ] RapidJson.DocumentParse_CrtAllocator_SSE42
[       OK ] RapidJson.DocumentParse_CrtAllocator_SSE42 (806 ms)
[ RUN      ] RapidJson.DocumentParseEncodedInputStream_MemoryStream_SSE42
[       OK ] RapidJson.DocumentParseEncodedInputStream_MemoryStream_SSE42 (1593 ms)
[ RUN      ] RapidJson.DocumentParseAutoUTFInputStream_MemoryStream_SSE42
[       OK ] RapidJson.DocumentParseAutoUTFInputStream_MemoryStream_SSE42 (2386 ms)
[ RUN      ] RapidJson.DocumentTraverse
[       OK ] RapidJson.DocumentTraverse (15 ms)
[ RUN      ] RapidJson.DocumentAccept
[       OK ] RapidJson.DocumentAccept (16 ms)


With RAPIDJSON_48BITPOINTER_OPTIMIZATION=1
[ RUN      ] RapidJson.DocumentParseInsitu_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParseInsitu_MemoryPoolAllocator_SSE42 (412 ms)
[ RUN      ] RapidJson.DocumentParseIterativeInsitu_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParseIterativeInsitu_MemoryPoolAllocator_SSE42 (428 ms)
[ RUN      ] RapidJson.DocumentParse_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParse_MemoryPoolAllocator_SSE42 (446 ms)
[ RUN      ] RapidJson.DocumentParseIterative_MemoryPoolAllocator_SSE42
[       OK ] RapidJson.DocumentParseIterative_MemoryPoolAllocator_SSE42 (447 ms)
[ RUN      ] RapidJson.DocumentParse_CrtAllocator_SSE42
[       OK ] RapidJson.DocumentParse_CrtAllocator_SSE42 (743 ms)
[ RUN      ] RapidJson.DocumentParseEncodedInputStream_MemoryStream_SSE42
[       OK ] RapidJson.DocumentParseEncodedInputStream_MemoryStream_SSE42 (1507 ms)
[ RUN      ] RapidJson.DocumentParseAutoUTFInputStream_MemoryStream_SSE42
[       OK ] RapidJson.DocumentParseAutoUTFInputStream_MemoryStream_SSE42 (2161 ms)
[ RUN      ] RapidJson.DocumentTraverse
[       OK ] RapidJson.DocumentTraverse (12 ms)
[ RUN      ] RapidJson.DocumentAccept
[       OK ] RapidJson.DocumentAccept (13 ms)
~~~

This PR also reduces `flags` in `GenericValue` from 32-bit to 16-bit and rearrange the layout. This will also benefit for other platforms, in which the `ShortString` can store more characters than before.